api/requirements.txt
fastapi
strawberry-graphql
cassandra-driver
cassandra-sigv4
mangum
pydantic
pydantic-core
